net, libs: Move NAD ref change helper
Move update_nad_references to libs/net/nad_ref.py so it is
available to all test packages.

Update all callers in l2_bridge nad_ref_change tests to import
from the new location.
