+++ tests/network/l2_bridge/nad_ref_change/lib_helpers.py
@@ -1,9 +1,7 @@
-from copy import deepcopy
 from typing import Final
 
 from kubernetes.dynamic import DynamicClient
 
-from libs.net.vmspec import wait_for_no_vmi_condition, wait_for_vmi_condition_status
 from libs.vm.factory import base_vmspec, fedora_vm
 from libs.vm.spec import (
     CloudInitNoCloud,
@@ -73,26 +71,6 @@ def assert_no_connectivity(
         server_bind_dev=server_bind_dev,
         expect_connectivity=False,
     )
-
-
-def update_nad_references(vm: BaseVirtualMachine, nad_name_by_net: dict[str, str]) -> None:
-    """Update secondary network NAD references and wait for the change to be fully applied.
-
-    Patches the VM spec atomically, then waits for the MigrationRequired condition to
-    appear (change detected) and disappear (migration completed).
-
-    Args:
-        vm: The virtual machine to update.
-        nad_name_by_net: Mapping of interface name to new NAD name.
-    """
-    resource_version = vm.vmi.instance.metadata.resourceVersion
-    networks = deepcopy(vm.template_spec.networks) or []
-    for network in networks:
-        if network.name in nad_name_by_net and network.multus:
-            network.multus.networkName = nad_name_by_net[network.name]
-    vm.set_networks(networks=networks)
-    wait_for_vmi_condition_status(vm=vm, condition="MigrationRequired", resource_version=resource_version)
-    wait_for_no_vmi_condition(vm=vm, condition="MigrationRequired")
 
 
 def two_secondary_bridge_vm(
